packagefunctions/requirements.txt
pandas
spacy
nltk
openai
gensim
SPARQLWrapper
wikipedia
python-dotenv
youtube_transcript_api
google-cloud
google-cloud-language
google-auth
google-api-core
google
google-cloud-bigquery
pandas-gbq
convertapi
requests
python-multipart
google-cloud-vision
moviepy
google-cloud-speech
google-cloud-storage
google-speech
BeautifulSoup4
numpy
regex
#string
